Rename and update structure of alpaka code and data-structure folders
