+++ benchmark/dataset_size/main.cpp
@@ -10,8 +10,8 @@
 #include <vector>
 
 #include "CLUEstering/CLUEstering.hpp"
-#include "CLUEstering/DataFormats/PointsHost.hpp"
-#include "CLUEstering/DataFormats/PointsDevice.hpp"
+#include "CLUEstering/data_structures/PointsHost.hpp"
+#include "CLUEstering/data_structures/PointsDevice.hpp"
 
 #include "utils/generation.hpp"
