+++ include/CLUEstering/data_structures/PointsHost.hpp
@@ -1,8 +1,8 @@
 
 #pragma once
 
-#include "CLUEstering/AlpakaCore/alpakaMemory.hpp"
-#include "CLUEstering/DataFormats/Common.hpp"
+#include "CLUEstering/data_structures/Common.hpp"
+#include "CLUEstering/internal/alpaka/memory.hpp"
 
 #include <alpaka/alpaka.hpp>
 #include <optional>
