+++ include/CLUEstering/data_structures/alpaka/SearchBox.hpp
