+++ include/CLUEstering/internal/alpaka/caching_allocator/get_device_caching_allocator.hpp
@@ -5,10 +5,10 @@
 #include <mutex>
 #include <vector>
 
-#include "CLUEstering/AlpakaCore/AllocatorConfig.hpp"
-#include "CLUEstering/AlpakaCore/CachingAllocator.hpp"
-#include "CLUEstering/AlpakaCore/alpakaDevices.hpp"
-#include "CLUEstering/AlpakaCore/getDeviceIndex.hpp"
+#include "CLUEstering/internal/alpaka/caching_allocator/allocator_config.hpp"
+#include "CLUEstering/internal/alpaka/caching_allocator/caching_allocator.hpp"
+#include "CLUEstering/internal/alpaka/devices.hpp"
+#include "CLUEstering/internal/alpaka/get_device_index.hpp"
 
 namespace clue {
