+++ tests/test_device_points.cpp
@@ -152,7 +152,8 @@ TEST_CASE("Test extrema functions on device points column") {
   clue::PointsDevice<2, Device> d_points(queue, size);
   clue::copyToDevice(queue, d_points, h_points);
 
-  auto max_it = clue::internal::algorithm::max_element(d_points.weight().begin(), d_points.weight().end());
+  auto max_it =
+      clue::internal::algorithm::max_element(d_points.weight().begin(), d_points.weight().end());
   auto max = 0.f;
   alpaka::memcpy(
       queue, clue::make_host_view(max), clue::make_device_view(alpaka::getDev(queue), *max_it));
@@ -174,5 +175,6 @@ TEST_CASE("Test reduction of device points column") {
   clue::PointsDevice<2, Device> d_points(queue, size);
   clue::copyToDevice(queue, d_points, h_points);
 
-  CHECK(clue::internal::algorithm::reduce(d_points.weight().begin(), d_points.weight().end()) == 499500.0f);
+  CHECK(clue::internal::algorithm::reduce(d_points.weight().begin(), d_points.weight().end()) ==
+        499500.0f);
 }
